Eksport funkcji utility z data_loader.c do utils.c
